# Improve perf with level guard in slogWrapper calls